+++ pkgs/dart_mcp/lib/src/server/elicitation_request_support.dart
@@ -5,12 +5,12 @@ part of 'server.dart';
 base mixin ElicitationRequestSupport on LoggingSupport {
   /// Whether or not the connected client supports elicitation.
   ///
-  /// Only safe to call after calling [initialize] on `super` since this is
-  /// based on the client capabilities.
+  /// Only safe to call after calling [initializeLegacy] on `super` since this
+  /// is based on the client capabilities.
   bool get supportsElicitation => clientCapabilities.elicitation != null;
 
   @override
-  FutureOr<InitializeResult> initialize(InitializeRequest request) {
+  FutureOr<InitializeResult> initializeLegacy(InitializeRequest request) {
     initialized.then((_) {
       if (!supportsElicitation) {
         log(
@@ -20,7 +20,7 @@ base mixin ElicitationRequestSupport on LoggingSupport {
         );
       }
     });
-    return super.initialize(request);
+    return super.initializeLegacy(request);
   }
 
   /// Sends an `elicitation/create` request to the client.