+++ pkgs/dart_mcp/lib/src/server/logging_support.dart
@@ -12,11 +12,12 @@ base mixin LoggingSupport on MCPServer {
   LoggingLevel loggingLevel = LoggingLevel.warning;
 
   @override
-  FutureOr<InitializeResult> initialize(InitializeRequest request) async {
+  FutureOr<ServerCapabilities> initialize(
+    ClientCapabilities clientCapabilities,
+  ) async {
     registerRequestHandler(SetLevelRequest.methodName, handleSetLevel);
 
-    return (await super.initialize(request))
-      ..capabilities.logging ??= Logging();
+    return (await super.initialize(clientCapabilities))..logging ??= Logging();
   }
 
   /// Sends a [LoggingMessageNotification] to the client, if the [loggingLevel]